+++ app/models/content_text.rb
@@ -1,7 +1,15 @@
 class ContentText < ApplicationRecord
   include Concerns::ContentType
 
+  # This triggers the searchable instance's after_save callback, which
+  # in turn updates all search indices.
+  after_save { parent.save! if linked_to_searchable_parent? }
+
   def content_type
     :text
+  end
+
+  def linked_to_searchable_parent?
+    column.present? && parent.present? && parent.searchable?
   end
 end