+++ app/models/search_index.rb
@@ -1,0 +1,5 @@
+class SearchIndex < ApplicationRecord
+  belongs_to :searchable, polymorphic: true
+
+  validates :locale, :name, presence: true
+end