+++ spec/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20161201104929) do
+ActiveRecord::Schema.define(version: 20170202170310) do
 
   create_table "addresses", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
     t.integer  "addressable_id"
@@ -248,6 +248,37 @@ ActiveRecord::Schema.define(version: 20161201104929) do
     t.datetime "created_at",      null: false
     t.datetime "updated_at",      null: false
     t.index ["source_uri"], name: "index_redirects_on_source_uri", using: :btree
+  end
+
+  create_table "search_indices", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string   "searchable_type"
+    t.integer  "searchable_id"
+    t.string   "locale"
+    t.string   "name"
+    t.text     "value",           limit: 65535
+    t.datetime "created_at",                    null: false
+    t.datetime "updated_at",                    null: false
+    t.index ["locale", "name"], name: "index_search_indices_on_locale_and_name", using: :btree
+    t.index ["searchable_type", "searchable_id"], name: "index_search_indices_on_searchable_type_and_searchable_id", using: :btree
+  end
+
+  create_table "search_modules", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string   "name"
+    t.boolean  "searchable"
+    t.integer  "weight"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["name", "searchable"], name: "index_search_modules_on_name_and_searchable", using: :btree
+  end
+
+  create_table "search_synonyms", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string   "locale"
+    t.string   "term"
+    t.text     "synonyms",   limit: 65535
+    t.datetime "created_at",               null: false
+    t.datetime "updated_at",               null: false
+    t.index ["locale"], name: "index_search_synonyms_on_locale", using: :btree
+    t.index ["term"], name: "index_search_synonyms_on_term", using: :btree
   end
 
   create_table "settings", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|